requirements.txt
# Web Scraping
newspaper3k==0.2.8
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
lxml_html_clean==0.1.1

# Google Gemini (free tier)
google-generativeai==0.5.4

# Email
secure-smtplib==0.1.1

# Scheduling
schedule==1.2.1

# Utilities
python-dotenv==1.0.0
feedparser==6.0.11
httpx==0.27.0
tqdm==4.66.2
tzdata==2024.1